Added some tests that expose problems with deeply nested states
This test case has a machine with four nested levels, A, B, C, D,
which happens to be the initial state.  A has two sibling states DONE
and FAIL, and the nodes A, B, C and D all react to A_EVENT, B_EVENT
and so on.

Currently the test only checks that the right exit handlers are
called, and exposes problems with handling deeply nested states.