Ensure that each orthogonal region is only entered once
If a transition has or resolves to multiple targets, and any of those
happen to target the same orthogonal region, an exception will be
thrown.  This is also checked recursively so that if an orthogonal
region is inside another orthogonal region, the parent orthogonal
region won't allow another transition to any other substate.