Failing test that exposes #106
The machine has orthogonal regions A, B, C.

A and B react to the event E by transitioning, and when this happens,
they raise two separate events, INT1 and INT2.

State C is sitting waiting on INT1 and INT2, and handles them by
trying to get to state C4, indicating that both INT1 and INT2 were
indeed attempted (in any order).

The test fails with

    AssertionError: expected
        { A: 'A2', B: 'B2', C: 'C2' } to deeply equal
        { A: 'A2', B: 'B2', C: 'C4' }

It sits in C2 instead of C4.