Guards can now access the interim states
When a transition occurs, and you want to check e.g. if you're in
states X and Y at the same time, and only perform e.g. an automatic
transition when you're in that state, you need to access the states
that are happening as part of a microstep.

The guard function now support a third argument, namely the "interim
state" so to speak, which is a StateValue suitable for using with
matchesState in order to check if you're "in" various states.