Added new EventProcess class to handle event processing
The new class will make sure that events are processed one at a time and in
the order they are met. Added some more transition tests besides the tests for
new EventProcess class. Reverted the previous partial implementation of
handling events synchronously with queue.

Please note that this commit changes the contract of the send interface for the
Interpreter class. The previous implementation returned the next state of the
state machine after processing the provided event regardless how many extra
events there was. The new implementation returns the current state regardless
of how many events it took to process.