Revises description of sub-state definition
I had difficulties to digest the paragraph about nested state nodes.
`Substate` is hard to understand for non native english speakers and can be easily misread as `substrate` or the like. `Sub-` as prefix should be easier to read.
Recursion means basically `running back` from its latin meaning. 
This has nothing to do with recursion. It's just a hierachical tree-like structure.
Maybe stating this as `nested state nodes` would also be fine or describing it as a `tree-like structure`.