Fix for “setCurrent is not a function” in @xstate/react for immediate mode.
Prior to this change, if the `immediate: true` option was supplied and the service had an transient transition in its initial state on `start()`, the `onTransition()` handler’s reference to `setCurrent` was `undefined` because `useState()` had not yet been executed to populate it.

Resolved this by splitting up the service creation logic such that the `onTransition()` handler is not added until after `useState()` has been executed and `setCurrent` is defined.