fix(xstate-react): `immediate` option with transient transitions on initial states (#756)
* Fix for “setCurrent is not a function” in @xstate/react for immediate mode.

Prior to this change, if the `immediate: true` option was supplied and the service had an transient transition in its initial state on `start()`, the `onTransition()` handler’s reference to `setCurrent` was `undefined` because `useState()` had not yet been executed to populate it.

Resolved this by splitting up the service creation logic such that the `onTransition()` handler is not added until after `useState()` has been executed and `setCurrent` is defined.

* Fixed tsc error TS2531: Object is possibly ‘null’

Functionally equivalent logic, but apparently more to the liking of TypeScript’s static analyzer? ¯\_(ツ)_/¯

* Reverted changes in favor of a simpler change

Reordered statements to execute immediate-mode’s `service.start()` after `useState()` call.

* Added unit tests for useMachine() immediate option

* Fixed incorrect unit test

Bad copy/paste - forgot to update `initial` in the machine for this test and didn’t actually test what I intended.