Add `types` field to all package.json files (#931)
* Add `types` field to all package.json's

IntelliJ based IDE's have difficulty trying to find the types unless you use the main path directly. I.E "@xstate/react/lib/index"

This is most likely down to the wording on the typescript publish page. (https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html)

It says "Also note that if your main declaration file is named index.d.ts and lives at the root of the package (next to index.js) you do not need to mark the "types" property, though it is advisable to do so.", so given that the types aren't at the root it isn't resolving it.

Adding this in fixes the issue.

* Remove packages marked as private

* Only publish `@xstate/text` in this changeset