Suggestion: making example valid in TypeScript
The original documented way to increment the simulated clock would not work in TypeScript because `service.clock` is of type `Clock` while only `SimulatedClock` has an `increment` method.

At first I thought I'd just use a type assertion like:
```
(service.clock as SimulatedClock).increment(1000);
```
and that may be good enough for the example.

I don't like unnecessary type assertions, so this PR proposes to keep the `SimulatedClock` instance in a local variable and increment that instead.

As I said, this is just a suggestion. What do you think?