Remove erroneous sub-category in Vue Recipe TOC
Prior to this change, the Vue recipe when navigated showed a repeated
sub category in the TOC. This was due to the top header being an h2
instead of an h1 like the other Doc files.

This change makes it an h1 to match and also run s the pre-commit
linters on the file.