docs(xstate-test): add note about nested states
This tripped me up before and caused me to spend a few hours debugging @xstate/test and my tests to determine why it seemed like tests were being run for the wrong state.

Turns out that @xstate/test executes not just the `meta.test` for the current state, but also all the parent states of the current state. Seems like a good thing to note in the docs as it's not immediately obvious (but makes sense now in retrospect).