Clarify Invoking with Context documentation
I was unclear from reading the documentation what would happen to the default context for a machine when invoking it with a `data` function. Would it merge the contexts together or would it replace it?

From testing, it appears to replace the context of the machine with what you return from the `data` function. This seems worth noting in the documentation