Let generics be inferred for `assign()` in fsm
The need to pass in the Context as a generic parameter in `assign()`
(when using `PropertyAssigner`) feels redundant when we already give
the Context to `createMachine` trough `StateMachine.Config`.

This commit changes `assign()` so that the type of the `assignment`
parameter in `assign()` is inferred, and then `TContext` is also
inferred, and so we don't need an explicit generic parameter when
using `PropertyAssigner`.

I also created a new type `Assignment` which is the union of
`Assigner` and `PropertyAssigner` to make the new code in `assign` a
little bit easier to read.