Doc: Add doc to useInterpret() in index.md
To clarify that `useInterpret` works with component lifecycles (like `useMachine` does), and how it compares to `useMachine`.
Had to dig into #1915 and aea4a2f to figure out the difference between `useInterpret` and `useMachine`. Realised the only difference was the observer, since `useInterpret` was previously named more clearly as `useMachineObserver`.